extend intToStr to accept long long

+++ src/base.cpp
@@ -139,7 +139,7 @@ namespace dss {
     return sstream.str();
   } // doubleToString
 
-  std::string intToString(const int _int, bool _hex) {
+  std::string intToString(const long long _int, bool _hex) {
     std::stringstream sstream;
     if(_hex) {
       sstream << std::hex << "0x";
